+++ rust/exo_rs/Cargo.toml
@@ -37,11 +37,6 @@ pyo3-log.workspace = true
 
 pidfile-rs = { git = "https://github.com/AndreiCravtov/pidfile-rs" }
 
-# macro dependencies
-extend = { workspace = true }
-delegate = { workspace = true }
-thiserror = "2.0"
-
 # async runtime
 tokio = { workspace = true, features = ["full"] }
 futures-lite.workspace = true
@@ -56,4 +51,3 @@ zenoh.workspace = true
 rand.workspace = true
 serde_json.workspace = true
 parking_lot.workspace = true
-pidfile-rs.workspace = true
